2024 November 10.11.2024 11.43am - Added from home

+++ inc/minirt.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   minirt.h                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: atang <atang@student.42.fr>                +#+  +:+       +#+        */
+/*   By: sentry <sentry@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/08/03 14:11:49 by atang             #+#    #+#             */
-/*   Updated: 2024/11/03 18:26:28 by atang            ###   ########.fr       */
+/*   Updated: 2024/11/10 09:41:06 by sentry           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -85,7 +85,7 @@ typedef struct s_Camera
 {
 	t_Vector3	position;
 	t_Vector3	orientation;
-	float		fov;
+	int			fov;
 }	t_Camera;
 
 typedef struct s_Light
